finish Hw v.1

+++ flask_orm/rooms/structures.py
@@ -4,5 +4,5 @@ room_structure = {
     "number": fields.Integer,
     "level": fields.Integer,
     "status": fields.String(40),
-    "price": fields.Integer
+    "price": fields.Float
 }
